requirements.txt
pyOpenSSL
requests